added support of tagged stream input

+++ lib/interleaver_impl.cc
@@ -96,8 +96,8 @@ namespace gr
 
       // nitems_to_process = std::min(nitems_to_process)
       // handle the first interleaved block special case
-      uint8_t cw_len = 4 + ((cw_cnt < m_sf - 2) ? 4 : m_cr);
-      uint8_t sf_app = ((cw_cnt < m_sf - 2) ||m_ldro) ? m_sf - 2 : m_sf;
+      uint8_t cw_len = 4 + (((int)cw_cnt < m_sf - 2) ? 4 : m_cr);
+      uint8_t sf_app = (((int)cw_cnt < m_sf - 2) ||m_ldro) ? m_sf - 2 : m_sf;
 
       nitems_to_process = std::min(nitems_to_process,(int)sf_app);
       if(std::floor((float)noutput_items/cw_len)==0)
@@ -105,7 +105,7 @@ namespace gr
         return 0;
       }
 
-      if (nitems_to_process >= sf_app || cw_cnt + nitems_to_process == m_frame_len)
+      if (nitems_to_process >= sf_app || cw_cnt + nitems_to_process == (uint32_t)m_frame_len)
       {        
         //propagate tag
         if(!cw_cnt)
@@ -146,7 +146,7 @@ namespace gr
             inter_bin[i][j] = cw_bin[mod((i - j - 1), sf_app)][i];
           }
           //For the first bloc we add a parity bit and a zero in the end of the lora symbol(reduced rate)
-          if ((cw_cnt == m_sf - 2)||m_ldro)
+          if (((int)cw_cnt == m_sf - 2)||m_ldro)
             inter_bin[i][sf_app] = accumulate(inter_bin[i].begin(), inter_bin[i].end(), 0) % 2;
 
           out[i] = bool2int(inter_bin[i]);
